Return documents
The search engine can now take in a search query and return documents containing this query. At the moment, it is only possible to search for a single word since the search engine just searches through the index and tries to find a perfect match. In addition, the found documents are returned in order of search term occurrence.