Add web crawler
Currently, the web crawler can retrieve HTML from a given origin path and extract all URLs that link to a different page and not to a section on the page. I'm using curl to retrieve the HTML document, and for extracting the URLs, I'm using gumbo. Additionally, I added a few tests to check if I can retrieve the HTML document, extract URLs, and the base URL.