(WIP) Exit threads
The problem was that the indexer_thread did not exit because it was still waiting for new documents to arrive, which meant it didn't check the flag. I tried to check the flag inside the queue 'wait' function, but that didn't work, and I don't know why. The current solution is to have an extra flag inside the queue, which the thread can check during waiting. I hope to find a better solution soon.